requirements.txt
appdirs==1.4.4
argcomplete==3.5.3
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
flask==3.0.3
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==2.1.5
pillow==10.4.0
pycups==2.0.4
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.15.1
python-dotenv==1.0.1
python-escpos==3.1
pyusb==1.2.1
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.3
six==1.17.0
typing-extensions==4.12.2
urllib3==2.2.3
waitress==3.0.0
werkzeug==3.0.6
zipp==3.20.2
